Splitt ut opprettBehandlingValg til egne react-komponenter
