+++ src/frontend/komponenter/Fagsak/Personlinje/Behandlingsmeny/OpprettBehandling/KravDatoFelt.tsx
@@ -1,14 +1,28 @@
 import React from 'react';
 
+import styled from 'styled-components';
+
+import { FamilieDatovelger } from '@navikt/familie-form-elements';
 import type { Felt } from '@navikt/familie-skjema';
 
 import type { FamilieIsoDate } from '../../../../../utils/kalender';
-import { FixedDatoVelger } from './OpprettBehandlingValg';
 
 interface IProps {
     kravMottattDato: Felt<FamilieIsoDate>;
     visFeilmeldinger: boolean;
 }
+
+const FixedDatoVelger = styled(FamilieDatovelger)`
+    .nav-datovelger__kalenderPortal__content {
+        position: fixed;
+    }
+
+    .nav-datovelger__kalenderknapp {
+        z-index: 0;
+    }
+
+    margin-top: 2rem;
+`;
 
 export const KravDatoFelt: React.FC<IProps> = ({ kravMottattDato, visFeilmeldinger }) => (
     <FixedDatoVelger
