--- src/frontend/komponenter/Fagsak/Personlinje/Behandlingsmeny/OpprettBehandling/OpprettBehandlingValg.tsx
@@ -1,196 +1,0 @@
-import React from 'react';
-
-import styled from 'styled-components';
-
-import { FamilieDatovelger, FamilieSelect } from '@navikt/familie-form-elements';
-import type { ISkjema } from '@navikt/familie-skjema';
-
-import { useApp } from '../../../../../context/AppContext';
-import type { ManuellJournalføringSkjemaFelter } from '../../../../../context/ManuellJournalførContext';
-import type { IBehandling } from '../../../../../typer/behandling';
-import { Behandlingstype, BehandlingÅrsak, behandlingÅrsak } from '../../../../../typer/behandling';
-import type { IMinimalFagsak } from '../../../../../typer/fagsak';
-import { Tilbakekrevingsbehandlingstype } from '../../../../../typer/tilbakekrevingsbehandling';
-import { ToggleNavn } from '../../../../../typer/toggles';
-import { hentAktivBehandlingPåMinimalFagsak } from '../../../../../utils/fagsak';
-import { BehandlingstemaSelect } from '../../../../Felleskomponenter/BehandlingstemaSelect';
-import type { VisningBehandling } from '../../../Saksoversikt/visningBehandling';
-import { kanOppretteFørstegangsbehandling, kanOppretteRevurdering } from './opprettBehandlingUtils';
-import type { IOpprettBehandlingSkjemaFelter } from './useOpprettBehandling';
-
-export const FixedDatoVelger = styled(FamilieDatovelger)`
-    .nav-datovelger__kalenderPortal__content {
-        position: fixed;
-    }
-
-    .nav-datovelger__kalenderknapp {
-        z-index: 0;
-    }
-
-    margin-top: 2rem;
-`;
-
-const StyledFamilieSelect = styled(FamilieSelect)`
-    label {
-        margin-top: 2rem;
-    }
-`;
-
-const StyledBehandlingstemaSelect = styled(BehandlingstemaSelect)`
-    label {
-        margin-top: 2rem;
-    }
-`;
-
-interface IProps {
-    opprettBehandlingSkjema:
-        | ISkjema<IOpprettBehandlingSkjemaFelter, IBehandling>
-        | ISkjema<ManuellJournalføringSkjemaFelter, string>;
-    minimalFagsak?: IMinimalFagsak;
-    erLesevisning?: boolean;
-    manuellJournalfør?: boolean;
-}
-
-interface BehandlingstypeSelect extends HTMLSelectElement {
-    value: Behandlingstype | '';
-}
-
-interface BehandlingÅrsakSelect extends HTMLSelectElement {
-    value: BehandlingÅrsak | '';
-}
-
-const OpprettBehandlingValg: React.FC<IProps> = ({
-    opprettBehandlingSkjema,
-    minimalFagsak,
-    erLesevisning = false,
-    manuellJournalfør = false,
-}) => {
-    const { toggles } = useApp();
-
-    const aktivBehandling: VisningBehandling | undefined = minimalFagsak
-        ? hentAktivBehandlingPåMinimalFagsak(minimalFagsak)
-        : undefined;
-
-    const kanOppretteTekniskEndring =
-        kanOppretteRevurdering(minimalFagsak, aktivBehandling) &&
-        toggles[ToggleNavn.kanBehandleTekniskEndring];
-
-    const kanOppretteTilbakekreving = !manuellJournalfør;
-
-    const visFeilmeldinger = opprettBehandlingSkjema.visFeilmeldinger;
-
-    const { behandlingsårsak, behandlingstype, behandlingstema } = opprettBehandlingSkjema.felter;
-
-    return (
-        <>
-            <FamilieSelect
-                {...behandlingstype.hentNavBaseSkjemaProps(
-                    opprettBehandlingSkjema.visFeilmeldinger
-                )}
-                erLesevisning={erLesevisning}
-                name={'Behandling'}
-                label={'Velg type behandling'}
-                onChange={(event: React.ChangeEvent<BehandlingstypeSelect>): void => {
-                    behandlingstype.onChange(event.target.value);
-                }}
-            >
-                <option disabled={true} value={''}>
-                    Velg
-                </option>
-                {kanOppretteFørstegangsbehandling(minimalFagsak, aktivBehandling) && (
-                    <option
-                        aria-selected={
-                            behandlingstype.verdi === Behandlingstype.FØRSTEGANGSBEHANDLING
-                        }
-                        value={Behandlingstype.FØRSTEGANGSBEHANDLING}
-                    >
-                        Førstegangsbehandling
-                    </option>
-                )}
-                {kanOppretteRevurdering(minimalFagsak, aktivBehandling) && (
-                    <option
-                        aria-selected={behandlingstype.verdi === Behandlingstype.REVURDERING}
-                        value={Behandlingstype.REVURDERING}
-                    >
-                        Revurdering
-                    </option>
-                )}
-
-                {kanOppretteTekniskEndring && (
-                    <option
-                        aria-selected={behandlingstype.verdi === Behandlingstype.TEKNISK_ENDRING}
-                        value={Behandlingstype.TEKNISK_ENDRING}
-                    >
-                        Teknisk endring
-                    </option>
-                )}
-
-                <option
-                    aria-selected={
-                        behandlingstype.verdi === Tilbakekrevingsbehandlingstype.TILBAKEKREVING
-                    }
-                    disabled={!kanOppretteTilbakekreving}
-                    value={Tilbakekrevingsbehandlingstype.TILBAKEKREVING}
-                >
-                    Tilbakekreving
-                </option>
-
-                {toggles[ToggleNavn.kanBehandleKlage] && (
-                    <option
-                        aria-selected={behandlingstype.verdi === Behandlingstype.KLAGE}
-                        value={Behandlingstype.KLAGE}
-                    >
-                        Klage
-                    </option>
-                )}
-            </FamilieSelect>
-
-            {behandlingsårsak.erSynlig && (
-                <StyledFamilieSelect
-                    {...behandlingsårsak.hentNavBaseSkjemaProps(visFeilmeldinger)}
-                    erLesevisning={erLesevisning}
-                    name={'Behandlingsårsak'}
-                    label={'Velg årsak'}
-                    onChange={(event: React.ChangeEvent<BehandlingÅrsakSelect>): void => {
-                        behandlingsårsak.onChange(event.target.value);
-                    }}
-                >
-                    <option disabled={true} value={''}>
-                        Velg
-                    </option>
-                    {Object.values(BehandlingÅrsak)
-                        .filter(
-                            årsak =>
-                                årsak !== BehandlingÅrsak.SATSENDRING &&
-                                årsak !== BehandlingÅrsak.TEKNISK_ENDRING &&
-                                (årsak !== BehandlingÅrsak.KORREKSJON_VEDTAKSBREV ||
-                                    toggles[ToggleNavn.kanManueltKorrigereMedVedtaksbrev])
-                        )
-                        .map(årsak => {
-                            return (
-                                <option
-                                    key={årsak}
-                                    aria-selected={behandlingsårsak.verdi === årsak}
-                                    value={årsak}
-                                >
-                                    {behandlingÅrsak[årsak]}
-                                </option>
-                            );
-                        })}
-                </StyledFamilieSelect>
-            )}
-
-            {behandlingstema.erSynlig && (
-                <StyledBehandlingstemaSelect
-                    behandlingstema={behandlingstema}
-                    erLesevisning={erLesevisning}
-                    visFeilmeldinger={opprettBehandlingSkjema.visFeilmeldinger}
-                    name="Behandlingstema"
-                    label="Velg behandlingstema"
-                />
-            )}
-        </>
-    );
-};
-
-export default OpprettBehandlingValg;
